requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altgraph==0.17.3
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
commonmark==0.9.1
DateTime==4.7
defusedxml==0.7.1
dj-database-url==0.5.0
dj-loguru==0.2.2
Django==4.1.5
django-bootstrap-v5==1.0.11
django-ckeditor==6.1.0
django-colorfield==0.8.0
django-cors-headers==3.13.0
django-js-asset==1.2.2
django-mptt==0.12.0
django-rest-framework==0.1.0
django-smart-selects==1.6.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environs==9.3.5
flake8==6.0.0
frozenlist==1.3.3
future==0.18.2
gunicorn==20.1.0
httpie==3.2.1
humanize==4.0.0
idna==3.4
isort==5.12.0
loguru==0.5.3
marshmallow==3.14.1
mccabe==0.7.0
multidict==6.0.3
packaging==23.0
pefile==2022.5.30
Pillow==9.4.0
psycopg2==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.13.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
PyJWT==2.6.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.7.1
pywin32-ctypes==0.2.0
requests==2.28.1
requests-toolbelt==0.10.1
rich==12.6.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
whitenoise==5.3.0
win32-setctime==1.1.0
yarl==1.8.2
zope.interface==5.5.2